+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Aded
+- Add `MetaLogger.metadata/0` function to return the logger metadata from the current process and
+  caller processes.
 
 ## [1.4.1] - 2021-06-21
 ### Changed